Create if Selection 1
Description

Note: You must use the values given and you must not "hard-code" results.

Write a Java program that checks to see if the age of a person is greater than or equal to 18. If it is, print out that the person can vote; otherwise, print that they cannot vote.
Use an integer variable age to represent the age of the person.


Sample Output(age=18)
Person can vote.

Sample Output(age=13)
Person cannot vote.