Create if Selection 2
Description

Note: You must use the values given and you must not "hard-code" results.

Write a java program that uses if selection statements to print out the cost of a toll based on the type of vehicle. Declare a character variable called vehicle. Your program should check the value of the character variable vehicle to see if it represents a car, a motorbike, a bus, a truck or a van. Depending on the type of vehicle it should print out the appropriate toll price. 

Vehicle representation is as follows: 'c' represents a car, 'm' represents a motorbike, 'b' represents a bus, 't' represents a truck and 'v' represents a van. 
Use lowercase case letters only - this is for grading purposes only.

The prices are as follows: Car - 2.00, Motorbike - 1.10, Bus - 4.25, Truck - 6.70 and Van - 4.00.



Sample Output

Motorbike - 1.10