Create Switch Selection 1
Description

Note: You must use the values given and you must not "hard-code" results.

Write a Java program that uses a switch statement to print out the month of the year given a value between 1 and 12, which is stored in an integer variable called month. For example, the value 1 would result in January being printed. If an integer that isn't a month is encountered it should output "Not a month".  Set month equal to 9 for evaluation.

Sample Output (month=9)
September

Sample Output (month=12)
December


Sample Output (month=13)
Not a month