Create While Loop 1
Description

Note: You must use the values given and you must not "hard-code" results.

Write a Java program that uses a while loop to find the factorial of a number. Declare an integer variable called num. 
Factorial (! denoted factorial in Maths) is the result of multiplying a sequence of natural numbers down to 1.

Examples of this are:

4! = 4 x 3 x 2 x 1 = 24

7! = 7 x 6 x 5 x 4 x 3 x 2 x 1 = 5040



The program should print in the format of the sample output below.



Sample Output
The factorial of 4 is 24