Create Backwards Sentence
Description

Write a Java program that takes a String, called sentence that references a lowercase String. The program should print the String in reverse word by word. For example, the string "pay no attention to that man behind the curtain" should be printed as follows:


curtain

the

behind

man

that

to

attention

no

pay

Sample Output 

curtain

the

behind

man

that

to

attention

no

pay